Allow to use PMUX when setting PMUX value
The syntax to set the PMUX value is:
PMUX [HARD] [POS] [AUX] <source> [<dest>]

The controller returns something like "PMUX POS AUX B5 R1" as read value
but doesn't allow to pass that string as set value.
Until now, "POS AUX B5 R1" was supposed to be used.

This change allows to pass as well the read value starting with "PMUX" (it will be ignored).